Implement PolyMouse!
IT WORKS! I ran my old Fitts' law test and got 4.3bits/s which
is the same as I got testing with my current mouse once.

It still glitches out occasionally, and can be a bit jumpy on mid-length moves.
The eye tracker also seems to disable itself after a while, possible when there's
no valid eyes for a while.